Logging a checkpoint here - not quite working but have a good structure going
