+++ .idea/misc.xml
@@ -4,6 +4,8 @@
   <component name="CidrRootsConfiguration">
     <sourceRoots>
       <file path="$PROJECT_DIR$" />
+      <file path="$PROJECT_DIR$/src" />
+      <file path="$PROJECT_DIR$/src/lib_src" />
     </sourceRoots>
     <libraryRoots>
       <file path="$PROJECT_DIR$/lib" />
