+++ CMakeLists.txt
@@ -29,3 +29,5 @@ add_custom_target(
     COMMAND platformio -c clion debug "$<$<NOT:$<CONFIG:All>>:-e${CMAKE_BUILD_TYPE}>"
     WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
 )
+
+add_executable(Z_DUMMY_TARGET ${SRC_LIST})
