+++ include/dac.h
@@ -4,9 +4,20 @@
 
 #ifndef AUDIO_SYNTH_DAC_H
 #define AUDIO_SYNTH_DAC_H
+/////////////////////////////////////////////////////////////////////////////////////////////
+////////////////////////////////////// DEFINES //////////////////////////////////////////////
+/////////////////////////////////////////////////////////////////////////////////////////////
 
 #define MCP4725_I2CADDR_DEFAULT (0x62) ///< Default i2c address
 
+/*******************************************************************************************/
 
+/////////////////////////////////////////////////////////////////////////////////////////////
+////////////////////////////////// FUNCTION PROTOS //////////////////////////////////////////
+/////////////////////////////////////////////////////////////////////////////////////////////
+
+void initDac(void);
+
+/*******************************************************************************************/
 
 #endif //AUDIO_SYNTH_DAC_H
