+++ include/defs.h
@@ -1,33 +1,26 @@
 #ifndef defs_h
 #define defs_h
 
+/////////////////////////////////////////////////////////////////////////////////////////////
+///////////////////////////////// FUNCTIONAL DEFINES ////////////////////////////////////////
+/////////////////////////////////////////////////////////////////////////////////////////////
+
 #define FALSE           0
 #define TRUE          !FALSE
 
+#define BitIsSet(VAR,Index)   (VAR & (1<<Index)) != 0
+#define BitIsNotSet(VAR,Index)  (VAR & (1<<Index)) == 0
+
+#define SetBit(VAR,Index)   VAR |= (1<<Index)
+#define ClearBit(VAR,Index)   VAR &= (uint8_t)(~(1<<Index))
+
 /////////////////////////////////////////////////////////////////////////////////////////////
-///////////////////////////////// PIN NUMBER DEFINES ////////////////////////////////////////
+/////////////////////////////////// PIN NUM DEFINES /////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////
 
-// Octave Shift Switches
-#define octDwn 11
-#define octUp 12
-
-// Primary output signal
-#define speakerOutPin   13
-
-// Wave select rotary input pins
-#define IN_SQUARE_SEL   22
-#define IN_TRI_SEL   23
-#define IN_SINE_SEL   24
-
-// Wave select gate driver output pins
-#define OUT_SQUARE_SEL  26
-#define OUT_TRI_SEL_1   27
-#define OUT_SINE_SEL_1  28
-#define OUT_SINE_SEL_2  29
-#define OUT_SINE_SEL_3  30
-#define OUT_SINE_SEL_4  31
-#define OUT_SINE_SEL_5  32
-
+#define IQS5xx_ADDR            0x74
+#define RDY_PIN               50
+#define RST_PIN               51
+#define END_WINDOW        (uint16_t)0xEEEE
 
 #endif
