+++ include/rtos.h
@@ -2,15 +2,14 @@
 // Created by Stephen Stranahan on 2/13/2022.
 //
 
-#ifndef AUDIO_SYNTH_OUTPUT_H
-#define AUDIO_SYNTH_OUTPUT_H
+#ifndef AUDIO_SYNTH_RTOS_H
+#define AUDIO_SYNTH_RTOS_H
 
 /////////////////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////// DEFINES //////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////
 
-// Primary output signal
-#define SPKR_OUT   13
+#define MCP4725_I2CADDR_DEFAULT (0x62) ///< Default i2c address
 
 /*******************************************************************************************/
 
@@ -18,10 +17,8 @@
 ////////////////////////////////// FUNCTION PROTOS //////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////
 
-void initOutput(void);
-
-void TaskProduceOutput( void *pvParameters );
+void initRtos(void);
 
 /*******************************************************************************************/
 
-#endif //AUDIO_SYNTH_OUTPUT_H
+#endif //AUDIO_SYNTH_RTOS_H
