+++ include/trackpad.h
@@ -5,4 +5,20 @@
 #ifndef AUDIO_SYNTH_TRACKPAD_H
 #define AUDIO_SYNTH_TRACKPAD_H
 
+/////////////////////////////////////////////////////////////////////////////////////////////
+////////////////////////////////////// DEFINES //////////////////////////////////////////////
+/////////////////////////////////////////////////////////////////////////////////////////////
+
+/*******************************************************************************************/
+
+/////////////////////////////////////////////////////////////////////////////////////////////
+////////////////////////////////// FUNCTION PROTOS //////////////////////////////////////////
+/////////////////////////////////////////////////////////////////////////////////////////////
+
+void TaskPollTrackpad( void *pvParameters );
+
+void initTrackpad(void);
+
+/*******************************************************************************************/
+
 #endif //AUDIO_SYNTH_TRACKPAD_H
