+++ platformio.ini
@@ -12,4 +12,6 @@
 platform = atmelavr
 board = megaatmega2560
 framework = arduino
-lib_deps = StackArray
+debug_tool = simavr
+lib_deps =
+    jdolinay/avr-debugger @ ~1.4
